refactor: ClockScene의 DefaultViewModel이 ViewModel protocol을 의존하도록 수정
