+++ Tempus/Presentation/ClockScene/Coordinator/DefaultClockCoordinator.swift
@@ -13,7 +13,7 @@ class DefaultClockCoordinator: ClockCoordinator {
     
     private weak var startApplyDelegate: StartApplyDelegate?
     
-    let clockViewModel: ClockViewModel
+    let clockViewModel: DefaultClockViewModel
     let clockViewController: ClockViewController
     
     init(startApplyDelegate: StartApplyDelegate) {
@@ -21,7 +21,7 @@ class DefaultClockCoordinator: ClockCoordinator {
         let tabBarSelectedImage = UIImage(systemName: "clock.fill")
         let tabBarItem = UITabBarItem(title: nil, image: tabBarImage, selectedImage: tabBarSelectedImage)
         
-        self.clockViewModel = ClockViewModel()
+        self.clockViewModel = DefaultClockViewModel()
         self.clockViewController = ClockViewController(viewModel: self.clockViewModel)
         self.clockViewController.tabBarItem = tabBarItem
         self.startApplyDelegate = startApplyDelegate
